some updates

+++ lib/shlinkedin_web/live/profile_live/show.ex
@@ -239,15 +239,17 @@ defmodule ShlinkedinWeb.ProfileLive.Show do
 
   @impl true
   def handle_event("delete-profile", %{"id" => id}, socket) do
-    profile = Profiles.get_profile_by_user_id(id)
-    {:ok, _} = Profiles.delete_profile(profile)
+    if Profiles.is_admin?(socket.assigns.profile) or socket.assigns.profile.id == id do
+      profile = Profiles.get_profile_by_user_id(id)
+      {:ok, _} = Profiles.delete_profile(profile)
 
-    {:noreply, socket |> fetch_posts()}
+      {:noreply, socket |> fetch_posts()}
+    end
   end
 
   @impl true
   def handle_event("delete", %{"id" => id}, socket) do
-    post = Timeline.get_post!(id)
+    {:ok, post} = Timeline.get_allowed_change_post(socket.assigns.profile, id)
     {:ok, _} = Timeline.delete_post(post)
 
     {:noreply, socket |> fetch_posts()}
@@ -255,7 +257,7 @@ defmodule ShlinkedinWeb.ProfileLive.Show do
 
   @impl true
   def handle_event("delete-comment", %{"id" => id}, socket) do
-    comment = Timeline.get_comment!(id)
+    {:ok, comment} = Timeline.get_allowed_change_comment(socket.assigns.profile, id)
     {:ok, _} = Timeline.delete_comment(comment)
 
     {:noreply,
